% Generated by hand
\name{calculate_baseline_emissions}
\alias{calculate_baseline_emissions}
\title{Calculate baseline emissions for AMS-I.J}
\usage{calculate_baseline_emissions(useful_output, baseline_emission_factor, output_col = "baseline_emissions_tco2e")}
\arguments{
  \item{useful_output}{Tibble produced by \code{calculate_useful_thermal_output()}.}
  \item{baseline_emission_factor}{Baseline emission factor in tCO2e/MWhth.}
  \item{output_col}{Name of the resulting emission column.}
}
\value{The input tibble with a baseline emission column in tCO2e.}
\description{Applies Equation (2) of AMS-I.J by multiplying useful thermal output by the baseline emission factor representing
displaced fossil fuel or electric water heating.}
\examples{
output <- tibble::tibble(useful_thermal_output_mwh = c(500, 620))
calculate_baseline_emissions(output, baseline_emission_factor = 0.25)
}
