% Generated by hand
\name{simulate_ams_ij_dataset}
\alias{simulate_ams_ij_dataset}
\title{Simulate a dataset compliant with AMS-I.J}
\usage{
simulate_ams_ij_dataset(
  n_sites = 10,
  n_periods = 12,
  start_year = 2023,
  start_month = 1,
  mean_useful_heat_mwh = 360,
  sd_useful_heat_mwh = 55,
  baseline_emission_factor = 0.22,
  auxiliary_emission_factor = 0.18,
  mean_auxiliary_share = 0.15,
  sd_auxiliary_share = 0.05
)
}
\arguments{
  \item{n_sites}{Number of solar water heating sites to simulate.}
  \item{n_periods}{Number of monitoring periods per site.}
  \item{start_year}{Calendar year for the first monitoring period.}
  \item{start_month}{Calendar month for the first monitoring period.}
  \item{mean_useful_heat_mwh}{Mean annual useful thermal output per site in MWhth.}
  \item{sd_useful_heat_mwh}{Standard deviation of annual useful thermal output in MWhth.}
  \item{baseline_emission_factor}{Baseline emission factor in tCO2e/MWhth.}
  \item{auxiliary_emission_factor}{Auxiliary system emission factor in tCO2e/MWhth.}
  \item{mean_auxiliary_share}{Mean share of auxiliary energy relative to useful solar heat.}
  \item{sd_auxiliary_share}{Standard deviation of the auxiliary share.}
}
\value{A tibble containing monitoring metadata, useful thermal output, auxiliary energy, and emission outcomes.}
\description{Generates a tidy dataset representing useful thermal output, auxiliary energy, and emission outcomes for solar water
heating systems.}
\examples{
simulate_ams_ij_dataset(n_sites = 3)
}
